+++ src/chrome-extension/content-script/components/save-page-message/components/save-page-content/save-note-editor/SaveNoteEditor.tsx
@@ -5,8 +5,9 @@ import {Node} from "slate";
 import {equals} from "ramda";
 import {defaultOptions} from "../../../../../../../common/rich-text-editor/options";
 import { defaultEditableProps } from "../../../../../../../common/rich-text-editor/editorFactory";
-import {useBasicPlugins} from "../../../../../../../common/rich-text-editor/base_plugins";
+import {useBasicPlugins} from "../../../../../../../common/rich-text-editor/plugins";
 import {useComponents} from "../../../../../../../common/rich-text-editor/components";
+import {NO_PLACEHOLDERS} from "../../../../../../../common/rich-text-editor/constants";
 
 export const SaveNoteEditor = (props: { initialValue: Node[], onChange: (newValue: Node[]) => void }) => {
     const { initialValue, onChange } = props
@@ -21,7 +22,7 @@ export const SaveNoteEditor = (props: { initialValue: Node[], onChange: (newValu
         <SlatePlugins
             id={"extensionNoteEditor"}
             plugins={useBasicPlugins()}
-            components={useComponents(false, false)}
+            components={useComponents(false, NO_PLACEHOLDERS)}
             onChange={updatePageContent}
             options={defaultOptions}
             editableProps={defaultEditableProps}
