+++ src/views/note-view-v2/NextGenNoteView.tsx
@@ -13,7 +13,8 @@ import {PeakEditor} from "../../common/rich-text-editor/editorFactory";
 import {wikiTitleEnforcer} from "../../common/rich-text-editor/editors/wiki/config";
 import {createPeakTitlePlugin} from "../../common/rich-text-editor/plugins/peak-title-plugin/PeakTitlePlugin";
 import {NoteTagSelect} from "../../common/rich-text-editor/plugins/peak-knowledge-plugin/components/peak-knowledge-node/peak-tag-select/component/NoteTagSelect";
-import {CaretLeftFilled} from "@ant-design/icons/lib";
+import {TITLE} from "../../common/rich-text-editor/types";
+import {DEFAULT_PLACEHOLDERS} from "../../common/rich-text-editor/constants";
 
 export const NextGenNoteView = (props: { note: PeakNote, selected_tags: PeakTag[] }) => {
     const { note, selected_tags } = props
@@ -62,8 +63,12 @@ export const NextGenNoteView = (props: { note: PeakNote, selected_tags: PeakTag[
                 getNodeContentSelectProps={getNodeContentSelectProps}
                 initialValue={noteContent}
                 currentPageId={currentPageId}
+                placeholderOverrides={[{
+                    key: TITLE,
+                    placeholder: 'Give your note a Title',
+                    hideOnBlur: false,
+                }]}
             />
         </div>
     )
-
 }
