+++ src/views/scratchpad/playground/Playground.tsx
@@ -54,8 +54,9 @@ import {
 import {defaultEditableProps} from "../../../common/rich-text-editor/editorFactory";
 import {optionsAutoformat} from "./playground-autoformat-rules";
 import {defaultOptions} from "../../../common/rich-text-editor/options";
-import {usePeakPlugins} from "../../../common/rich-text-editor/base_plugins";
+import {usePeakPlugins} from "../../../common/rich-text-editor/plugins";
 import {useComponents} from "../../../common/rich-text-editor/components";
+import {DEFAULT_PLACEHOLDERS} from "../../../common/rich-text-editor/constants";
 
 const id = 'Examples/Playground'
 
@@ -116,7 +117,7 @@ export const Plugins = () => {
             <SlatePlugins
                 id={id}
                 plugins={usePeakPlugins()}
-                components={useComponents()}
+                components={useComponents(true, DEFAULT_PLACEHOLDERS)}
                 options={defaultOptions}
                 editableProps={defaultEditableProps}
                 initialValue={initialValueHighlight}
